+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Logo Component.
+- Intro Component.
 
 ## [0.2.0] - 2021-01-28